Update serialize and deserialize methods.
The serialize and deserialize methods for this plugin were last updated prior to MAP-Client version 0.12.0 so are incompatible with the latest version of the MAP-Client. It is no longer necessary to serialize and deserialize each configuration entry explicitly, instead the serialization methods should utilize the Python dictionary class to automatically serialize and deserialize configuration state values. The configuration state class has been updated to make use of this functionality.